2024 Data Science Summer Internship
For the Data Science Internship for Xavier, we used data directly from the University's Registrar's Office. Our goal was to answer the questions of the Registrar's Office and all department chairs over a month. Through the course of this Internship, we were able to provide direct models of the Student's experience at Xavier University, including the requirements that took the most amount of semesters to complete, the most challenging requirements, the average GPA for each requirement, which categories of Majors had to take the most out of significant course to fulfill the core and many more. The product of our work has since then been the basis for the review of the Core Curriculum as we gave concrete evidence for how students were experiencing the core instead of reviewing the core off of word of mouth and consensus. It is now being used to construct a new core curriculum for the entire university, which will be rolled out in the near future.